refactor: move Masm component and Wat emission to the codegen and frontend
